integrations/requirements.txt
apache-airflow-providers-weaviate==1.3.4
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
weaviate-client==4.5.5
pyarrow>=10.0.0